Improve transaction types
